Add unified SettingsPanel with tabs
