--- cueit-admin/src/ConfigPanel.jsx
@@ -1,101 +1,0 @@
-import React from 'react';
-import axios from 'axios';
-
-export default function ConfigPanel({ open, onClose, config, setConfig }) {
-  const api = import.meta.env.VITE_API_URL;
-
-  const saveConfig = async () => {
-    try {
-      await axios.put(`${api}/api/config`, config);
-      alert('Saved');
-    } catch (err) {
-      alert('Failed to save configuration');
-    }
-  };
-
-  return (
-    <div
-      className={`fixed inset-0 bg-black/50 z-60 transition-opacity ${open ? 'opacity-100 pointer-events-auto' : 'opacity-0 pointer-events-none'}`}
-    >
-      <div
-        className={`absolute right-0 top-0 bottom-0 w-80 bg-gray-800 text-white p-6 transform transition-transform duration-300 ${open ? 'translate-x-0' : 'translate-x-full'}`}
-      >
-        <button onClick={onClose} className="mb-4 text-right w-full hover:text-gray-300">
-          ✖
-        </button>
-        <h2 className="text-lg mb-4">Configuration</h2>
-        <div className="space-y-3 text-sm">
-          <label className="block">
-            Logo URL
-            <input
-              type="text"
-              value={config.logoUrl || ''}
-              onChange={(e) => setConfig({ ...config, logoUrl: e.target.value })}
-              className="mt-1 w-full px-2 py-1 rounded text-black"
-            />
-          </label>
-          <label className="block">
-            Favicon URL
-            <input
-              type="text"
-              value={config.faviconUrl || ''}
-              onChange={(e) => setConfig({ ...config, faviconUrl: e.target.value })}
-              className="mt-1 w-full px-2 py-1 rounded text-black"
-            />
-          </label>
-          <label className="block">
-            Welcome Message
-            <input
-              type="text"
-              value={config.welcomeMessage || ''}
-              onChange={(e) => setConfig({ ...config, welcomeMessage: e.target.value })}
-              className="mt-1 w-full px-2 py-1 rounded text-black"
-            />
-          </label>
-          <label className="block">
-            Help Message
-            <input
-              type="text"
-              value={config.helpMessage || ''}
-              onChange={(e) => setConfig({ ...config, helpMessage: e.target.value })}
-              className="mt-1 w-full px-2 py-1 rounded text-black"
-            />
-          </label>
-          <button onClick={saveConfig} className="px-4 py-2 bg-blue-600 text-white rounded mt-2">
-            Save
-          </button>
-          <div className="pt-4 border-t border-gray-700 text-gray-300 text-xs space-y-2">
-            <div>Environment</div>
-            <label className="block">
-              API URL
-              <input
-                type="text"
-                value={import.meta.env.VITE_API_URL}
-                readOnly
-                className="mt-1 w-full px-2 py-1 rounded bg-gray-700 text-gray-400"
-              />
-            </label>
-            <label className="block">
-              Default Logo
-              <input
-                type="text"
-                value={import.meta.env.VITE_LOGO_URL}
-                readOnly
-                className="mt-1 w-full px-2 py-1 rounded bg-gray-700 text-gray-400"
-              />
-            </label>
-            <label className="block">
-              Default Favicon
-              <input
-                type="text"
-                value={import.meta.env.VITE_FAVICON_URL}
-                readOnly
-                className="mt-1 w-full px-2 py-1 rounded bg-gray-700 text-gray-400"
-              />
-            </label>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
