+++ cueit-admin/src/__tests__/App.test.jsx
@@ -39,7 +39,6 @@ describe('App filtering and config', () => {
     render(<App />);
     await screen.findByText('Alice');
     await userEvent.click(screen.getByLabelText('Settings'));
-    await userEvent.click(screen.getByText('Config'));
     const logoInput = screen.getByLabelText('Logo URL');
     await userEvent.clear(logoInput);
     await userEvent.type(logoInput, 'newlogo.png');
